Resampling approach for model evaluation

evaluate_resampling() uses repeated K-fold cross-validation and the Root Mean Square Error (RMSE) of testing sets to measure the predictive power of a single model. Methods are provided for trending::trending_model (and lists of these) objects.

evaluate_resampling(x, ...)

# S3 method for default
evaluate_resampling(x, ...)

# S3 method for trending_model
evaluate_resampling(
  x,
  data,
  metric = c("rmse", "rsq", "mae"),
  metric_arguments = list(na.rm = TRUE),
  v = nrow(data),
  repeats = 1,
  ...
)

# S3 method for list
evaluate_resampling(
  x,
  data,
  metric = c("rmse", "rsq", "mae"),
  metric_arguments = list(na.rm = TRUE),
  v = nrow(data),
  repeats = 1,
  ...
)

Arguments

x
An R object.

...
Not currently used.

data
a data.frame containing data (including the response variable and all predictors) used in the specified model.

metric
One of "rmse" (see calculate_rmse), "mae" (see calculate_mae) and "rsq" (see calculate_rsq).

metric_arguments
A named list of arguments passed to the underlying functions that calculate the metrics.

v
the number of equally sized data partitions to be used for K-fold cross-validation; v cross-validations will be performed, each using v - 1 partition as training set, and the remaining partition as testing set. Defaults to the number of row in data, so that the method uses leave-one-out cross validation, akin to Jackknife except that the testing set (and not the training set) is used to compute the fit statistics.

repeats
the number of times the random K-fold cross validation should be repeated for; defaults to 1; larger values are likely to yield more reliable / stable results, at the expense of computational time

Details

These functions wrap around existing functions from several packages. evaluate_resampling.trending_model() and evaluate_resampling.list() both use rsample::vfold_cv() for sampling and, for the calculating the different metrics, the yardstick package.

See also

Examples

x <- rnorm(100, mean = 0)
y <- rpois(n = 100, lambda = exp(x + 1))
dat <- data.frame(x = x, y = y)
model <- trending::glm_model(y ~ x, poisson)
models <- list(
  poisson_model = trending::glm_model(y ~ x, poisson),
  linear_model = trending::lm_model(y ~ x)
)

evaluate_resampling(model, dat)

#> # A tibble: 100 × 10
#>    metric result warnings errors fitted_model fitting_warnings fitting_errors
#>  * <chr>   <dbl> <list>   <list> <named list> <named list>     <named list>  
#>  1 rmse   3.29   <NULL>   <NULL> <glm>        <NULL>           <NULL>        
#>  2 rmse   0.0141 <NULL>   <NULL> <glm>        <NULL>           <NULL>        
#>  3 rmse   1.23   <NULL>   <NULL> <glm>        <NULL>           <NULL>        
#>  4 rmse   0.332  <NULL>   <NULL> <glm>        <NULL>           <NULL>        
#>  5 rmse   1.54   <NULL>   <NULL> <glm>        <NULL>           <NULL>        
#>  6 rmse   3.08   <NULL>   <NULL> <glm>        <NULL>           <NULL>        
#>  7 rmse   0.968  <NULL>   <NULL> <glm>        <NULL>           <NULL>        
#>  8 rmse   2.34   <NULL>   <NULL> <glm>        <NULL>           <NULL>        
#>  9 rmse   1.82   <NULL>   <NULL> <glm>        <NULL>           <NULL>        
#> 10 rmse   8.81   <NULL>   <NULL> <glm>        <NULL>           <NULL>        
#> # … with 90 more rows, and 3 more variables: pred <named list>,
#> #   predicting_warnings <named list>, predicting_errors <named list>

evaluate_resampling(models, dat)

#> # A tibble: 200 × 11
#>    model_name    metric result warnings errors fitted_model fitting_warnings
#>    <chr>         <chr>   <dbl> <list>   <list> <named list> <named list>    
#>  1 poisson_model rmse    0.897 <NULL>   <NULL> <glm>        <NULL>          
#>  2 poisson_model rmse    1.09  <NULL>   <NULL> <glm>        <NULL>          
#>  3 poisson_model rmse    1.39  <NULL>   <NULL> <glm>        <NULL>          
#>  4 poisson_model rmse    0.356 <NULL>   <NULL> <glm>        <NULL>          
#>  5 poisson_model rmse    0.671 <NULL>   <NULL> <glm>        <NULL>          
#>  6 poisson_model rmse    2.93  <NULL>   <NULL> <glm>        <NULL>          
#>  7 poisson_model rmse    0.162 <NULL>   <NULL> <glm>        <NULL>          
#>  8 poisson_model rmse    6.18  <NULL>   <NULL> <glm>        <NULL>          
#>  9 poisson_model rmse    0.717 <NULL>   <NULL> <glm>        <NULL>          
#> 10 poisson_model rmse    1.54  <NULL>   <NULL> <glm>        <NULL>          
#> # … with 190 more rows, and 4 more variables: fitting_errors <named list>,
#> #   pred <named list>, predicting_warnings <named list>,
#> #   predicting_errors <named list>